Remove gradle.properties to fix packaging
Found that this service's packaged Java code was being
pushed into build/distributions/consent-history-api.zip/consent-history-api/lib/, which causes AWS Lambda to fail to find the classes when
the service code is deployed.

The intermediate consent-history-api directory should not
be generated in the zip file.

After dropping the file, the JARs are correctly
generated in build/distributions/consent-history-api.zip/lib.